Abstract out live TI search functions to own component
