+++ client/components/App.jsx
@@ -5,6 +5,7 @@ import Prompt from './Prompt';
 import Setup from './Setup';
 import apiKey from '../../apiKey.js';
 import searchStaticTiHistory from './searchStaticTiHistory.js';
+import searchLiveTiHistory from './searchLiveTiHistory.js';
 
 class App extends Component {
   state = {
@@ -46,7 +47,7 @@ class App extends Component {
     });
   }
 
-  copyPrompt = async (event) => {
+  copyPrompt = async event => {
     this.setState({ promptSelected: true, suggestedPrompt: '' });
     const promptName = event.target.id;
     const promptId = {
@@ -149,45 +150,12 @@ class App extends Component {
   };
 
   getPrevInterviews = async email => {
-    const path = 'https://sheets.googleapis.com/v4/spreadsheets/' +
-      '1ObVQGqm894fzjeM5vcG2qysYKDng4g8rtgyIwS3vJh8' +
-      '?ranges=%27Form%20Responses%27&includeGridData=true';
-    const data = await gapi.client.request({ path });
-    const liveTiRows = this.processLiveTiData(data, email);
     const staticTiRows = searchStaticTiHistory(email);
+    const liveTiRows = await searchLiveTiHistory(email);
     const allRecords = (liveTiRows.join('') + staticTiRows.join('')).toLowerCase();
     const suggestedPrompt = !allRecords.includes('version') ? 'Version Control'
       : !allRecords.includes('mrp') ? 'MRP' : 'Book Library';
     this.setState({ staticTiRows, liveTiRows, suggestedPrompt });
-  };
-
-  formatLiveTiRow = row => {
-    const unwrappedValues = row.map(cell => cell.formattedValue);
-    const time = moment(unwrappedValues[0]).format('MMMM Do YYYY');
-    const status = unwrappedValues[5];
-    const prompt = unwrappedValues[10];
-    return [time, status, prompt].filter(el => el).join(', ');
-  };
-
-  processLiveTiData = (data, email) => {
-    const { result: { sheets: [sheet] } } = data;
-    const { data: [{ rowData : rows }]} = sheet;
-    // The sheet has a number of empty rows at the bottom,
-    // which .filter would needlessly iterate through.
-    const unformattedLiveTiRows = [];
-    for (let i = 0; i < rows.length; i++) {
-      const { values } = rows[i];
-      // Empty cells are empty objects, so we can't just check truthiness.
-      if (!Object.keys(values[0]).length) { // If we're past the filled rows
-        break;
-      }
-
-      const rowEmail = values[4].formattedValue;
-      if (rowEmail.trim() === email.trim()) { // No misleading whitespace
-        unformattedLiveTiRows.push(values);
-      }
-    }
-    return unformattedLiveTiRows.map(this.formatLiveTiRow);
   };
 
   handleLogin = (loggingIn = true) => {
